Optimize file structure
