+++ qlib/rl/order_execution/reward.py
@@ -9,7 +9,7 @@ import numpy as np
 
 from qlib.rl.reward import Reward
 
-from .simulator_simple import SAOEMetrics, SAOEState
+from qlib.rl.order_execution.state import SAOEMetrics, SAOEState
 
 __all__ = ["PAPenaltyReward"]
